feat: improve type inference model (#169)
Creates a state storage plugin for saving type information.

Type system improvements:
* disallow anonymous ids in relations/elements
* ensure function ids are included in function definition relations
* map function calls correctly to the original function definition
* remove notion of 'index' in array type
* differentiate between strong and weak relations between type models
* removed type enum from statement constructors
* always keep typeid in statements

+++ tools/javascript/lib/JavaScriptLauncher.ts
@@ -363,7 +363,9 @@ export class JavaScriptLauncher extends Launcher {
 
     const startTypeResolving = Date.now();
     JavaScriptLauncher.LOGGER.info("Extracting types");
-    this.rootContext.extractTypes();
+    this.rootContext.getAllElements();
+    this.rootContext.getAllRelations();
+    this.rootContext.getAllObjectTypes();
     JavaScriptLauncher.LOGGER.info("Resolving types");
     this.rootContext.resolveTypes();
     timeInMs = (Date.now() - startTypeResolving) / 1000;
@@ -828,7 +830,9 @@ export class JavaScriptLauncher extends Launcher {
 
   async exit(): Promise<void> {
     JavaScriptLauncher.LOGGER.info("Exiting");
-    this.runner.process.kill();
+    if (this.runner && this.runner.process) {
+      this.runner.process.kill();
+    }
     // TODO should be cleanup step in tool
     // Finish
     JavaScriptLauncher.LOGGER.info("Deleting temporary directories");
